fix: use event duration instead of whole suite result's duartion when making a summary
